Manually merged changes to trunk in commit 14331
Fixes the problem that when a harvest and termination operation occur on the same day, the stir report does not properly identify the termination operation. The fix for multiple end of season and hydrobal reports on the same day also broke the attaching of names to the harvest/termination operations in this scenario as well. This fixes that as well.

git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@14332 77af8a48-1994-4e4a-b6f4-b1c89642997b