Manual merge of changes from checkin 15081.  rpt_season_flag was already being properly initialized.
Mantis issue #1994

Initialization of rpt_season_flag was added in checkin 15076 and 15077. Changed the initialization value to true for proper functioning.

Added warning when the heat unit accumulation found when tring to grow a crop with a set days to maturity is not enough for vlaid crop growth. Used range between tbase and topt as an initial scaling guess.

Capped hui index to be a maximum of 1.0

git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@15082 77af8a48-1994-4e4a-b6f4-b1c89642997b