Manual merge of Trunk commit 15340 changes.
Height increase only occurs if there is a stem mass increase. This prevents a height anomoly if a crop which cannot regrow is cut in the early season. In old code, height will increase with no biomass increase occurring. SAI under this management was correct, but height was not. Ultimately, under this condition, the remaining crop biomass should become residue since the crop should be dead. Modeling the height of the gowing point should be modelled to enble a better determination of the regrowth potential of the crop at any stage of growth.

git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@15341 77af8a48-1994-4e4a-b6f4-b1c89642997b